requirements.txt
flask==3.1.2
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pymssql==2.3.13
python-dotenv==1.0.0